Moved sub routes added by all four members to separate files imported subroutes in server.js file
