+++ src/server/shopRoutes1.js
@@ -1,0 +1,24 @@
+/*
+* Added by: A.P.Manjari Mandara (IT18076844)
+* file: shopRoutes.js
+* description: routes related to shopping cart
+*
+* */
+
+
+
+
+//
+// let express = require("express")
+// let shopRoutes = express.Router();
+//
+// let shop = require('../schema/shop');
+//
+
+
+
+////
+
+//Manjari's part' here
+
+////
